# api.yml
openapi: "3.0.0"
info:
    version: 1.0.0
    title: Checkout Backend Service - IFP 2026
    description: |
        Backend service for e-commerce checkout system with dynamic promotions.
        Supports product management, shopping cart, promotions, and order processing.
    license:
        name: MIT
    contact:
        name: API Support
        email: support@example.com
servers:
    - url: http://localhost:8080/api
      description: Local development server
    - url: https://api.example.com/api
      description: Production server
...
paths:
...
    /cart:
        get:
            tags:
                - Cart
            summary: Get current cart
            description: Retrieve the active shopping cart with all items
            operationId: getCart
            responses:
                "200":
                    description: Cart retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CartResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        post:
            tags:
                - Cart
            summary: Add item to cart
            description: Add a product to the shopping cart or increase quantity if already exists
            operationId: addToCart
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AddToCartRequest"
            responses:
                "200":
                    description: Item added to cart successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CartResponse"
                "400":
                    description: Invalid request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Insufficient stock
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /cart/{itemId}:
        put:
            tags:
                - Cart
            summary: Update cart item quantity
            description: Update the quantity of a specific item in the cart
            operationId: updateCartItem
            parameters:
                - name: itemId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: Cart item ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateCartItemRequest"
            responses:
                "200":
                    description: Cart item updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CartResponse"
                "400":
                    description: Invalid request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Cart item not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Insufficient stock
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            tags:
                - Cart
            summary: Remove item from cart
            description: Remove a specific item from the shopping cart
            operationId: removeCartItem
            parameters:
                - name: itemId
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: Cart item ID
            responses:
                "200":
                    description: Cart item removed successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CartResponse"
                "404":
                    description: Cart item not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

    # ============================================
    # CHECKOUT ENDPOINTS
    # ============================================
    /checkout:
        post:
            tags:
                - Checkout
            summary: Process checkout
            description: Process the checkout, apply promotions, reduce stock, and create order
            operationId: checkout
            parameters:
                - name: Idempotency-Key
                  in: header
                  required: true
                  schema:
                      type: string
                  description: Unique key to prevent duplicate transactions
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CheckoutRequest"
            responses:
                "200":
                    description: Checkout processed successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CheckoutResponse"
                "400":
                    description: Invalid request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Cart or product not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: Duplicate transaction or insufficient stock
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "422":
                    description: Cart is empty or invalid
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /payment/confirm:
        post:
            tags:
                - Checkout
            summary: Confirm payment
            description: Confirm payment status for an order
            operationId: confirmPayment
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PaymentConfirmRequest"
            responses:
                "200":
                    description: Payment confirmed successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PaymentConfirmResponse"
                "400":
                    description: Invalid request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Order not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "500":
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

    # ============================================
    # ORDER ENDPOINTS
    # ============================================
...
components:
    schemas:
...
        CartItem:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                    description: Cart item ID
                product_sku:
                    type: string
                    description: Product SKU
                    example: "120P90"
                product_name:
                    type: string
                    description: Product name
                    example: "Google Home"
                quantity:
                    type: integer
                    minimum: 1
                    description: Quantity
                    example: 3
                unit_price:
                    type: number
                    format: float
                    minimum: 0
                    description: Unit price
                    example: 49.99
                total_price:
                    type: number
                    format: float
                    minimum: 0
                    description: Total price before discount
                    example: 149.97
                discount_amount:
                    type: number
                    format: float
                    minimum: 0
                    description: Discount amount applied
                    example: 49.99
                final_price:
                    type: number
                    format: float
                    minimum: 0
                    description: Final price after discount
                    example: 99.98
                promotion_applied:
                    type: string
                    description: Name of promotion applied (if any)
                    example: "Buy 3 Google Homes for price of 2"
        CartSummary:
            type: object
            properties:
                subtotal:
                    type: number
                    format: float
                    minimum: 0
                    description: Total before discounts
                    example: 5549.96
                total_discount:
                    type: number
                    format: float
                    minimum: 0
                    description: Total discount applied
                    example: 79.99
                grand_total:
                    type: number
                    format: float
                    minimum: 0
                    description: Final total after discounts
                    example: 5469.97
        CartResponse:
            type: object
            properties:
                cart_id:
                    type: string
                    format: uuid
                    description: Cart ID
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/CartItem"
                summary:
                    $ref: "#/components/schemas/CartSummary"
                status:
                    type: string
                    enum: [active, completed, cancelled]
                    description: Cart status
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
        AddToCartRequest:
            type: object
            required:
                - product_sku
                - quantity
            properties:
                product_sku:
                    type: string
                    description: Product SKU to add
                    example: "120P90"
                quantity:
                    type: integer
                    minimum: 1
                    description: Quantity to add
                    example: 2
        UpdateCartItemRequest:
            type: object
            required:
                - quantity
            properties:
                quantity:
                    type: integer
                    minimum: 1
                    description: New quantity
                    example: 5

        # ==========================================
        # CHECKOUT SCHEMAS
        # ==========================================
        CheckoutRequest:
            type: object
            required:
                - cart_id
                - user_id
                - payment_method
                - shipping_address
                - idempotency_key
            properties:
                cart_id:
                    type: string
                    format: uuid
                    description: Cart ID to checkout
                    example: "550e8400-e29b-41d4-a716-446655440000"
                user_id:
                    type: string
                    format: uuid
                    description: User ID
                    example: "660e8400-e29b-41d4-a716-446655440000"
                payment_method:
                    type: string
                    description: Payment method
                    example: "credit_card"
                shipping_address:
                    type: string
                    description: Shipping address
                    example: "Jl. Contoh No. 123, Jakarta, 12345, Indonesia"
                idempotency_key:
                    type: string
                    description: Unique key to prevent duplicate transactions
                    example: "unique-key-12345"
        PaymentConfirmRequest:
            type: object
            required:
                - order_id
                - payment_id
                - payment_status
                - transaction_reference
            properties:
                order_id:
                    type: string
                    format: uuid
                    description: Order ID
                    example: "770e8400-e29b-41d4-a716-446655440000"
                payment_id:
                    type: string
                    format: uuid
                    description: Payment ID
                    example: "880e8400-e29b-41d4-a716-446655440000"
                payment_status:
                    type: string
                    enum: [success, failed]
                    description: Payment status
                    example: "success"
                transaction_reference:
                    type: string
                    description: Transaction reference from payment gateway
                    example: "ref_1234567890"
        PaymentConfirmResponse:
            type: object
            properties:
                order_id:
                    type: string
                    format: uuid
                status:
                    type: string
                    example: "success"
                message:
                    type: string
                    example: "Payment confirmed successfully"

        # ==========================================
        # ORDER SCHEMAS
        # ==========================================
        OrderItem:
            type: object
            properties:
                sku:
                    type: string
                    example: "120P90"
                quantity:
                    type: integer
                    example: 3
                unit_price:
                    type: number
                    format: float
                    example: 49.99
                original_subtotal:
                    type: number
                    format: float
                    example: 149.97
                discount:
                    type: number
                    format: float
                    example: 49.99
                final_price:
                    type: number
                    format: float
                    example: 99.98
                promotion_applied:
                    type: string
                    example: "Buy 3 Google Homes for price of 2"
        PromotionBreakdown:
            type: object
            properties:
                promotion:
                    type: string
                    description: Promotion name
                    example: "Buy 3 Google Homes for price of 2"
                saving:
                    type: number
                    format: float
                    description: Total saving from this promotion
                    example: 49.99
                items_affected:
                    type: array
                    items:
                        type: string
                    description: SKUs affected by this promotion
                    example: ["120P90"]
        CheckoutSummary:
            type: object
            properties:
                subtotal:
                    type: number
                    format: float
                    example: 5579.96
                total_discount:
                    type: number
                    format: float
                    example: 79.99
                grand_total:
                    type: number
                    format: float
                    example: 5499.97
                promotion_breakdown:
                    type: array
                    items:
                        $ref: "#/components/schemas/PromotionBreakdown"
        PaymentInfo:
            type: object
            properties:
                payment_id:
                    type: string
                    format: uuid
                    example: "880e8400-e29b-41d4-a716-446655440000"
                status:
                    type: string
                    enum: [pending, success, failed]
                    example: "pending"
                amount:
                    type: number
                    format: float
                    example: 5499.97
                payment_method:
                    type: string
                    example: "credit_card"
                transaction_id:
                    type: string
                    example: "txn_12345678"
        CheckoutResponse:
            type: object
            properties:
                order_id:
                    type: string
                    format: uuid
                order_number:
                    type: string
                    example: "ORD-20260128-0001"
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/OrderItem"
                summary:
                    $ref: "#/components/schemas/CheckoutSummary"
                payment:
                    $ref: "#/components/schemas/PaymentInfo"
                shipping_address:
                    type: string
                    example: "Jl. Contoh No. 123, Jakarta, 12345, Indonesia"
                status:
                    type: string
                    enum: [pending_payment, paid, processing, shipped, delivered, completed, cancelled, payment_failed]
                    example: "pending_payment"
                created_at:
                    type: string
                    format: date-time

        # ==========================================
        # ORDER LIST SCHEMAS
        # ==========================================
...
        ErrorResponse:
            type: object
            properties:
                status:
                    type: string
                    example: "error"
                message:
                    type: string
                    example: "Invalid request parameters"
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-API-Key

# Optional: Add security if needed
# security:
#   - ApiKeyAuth: []